requirements.txt
fastapi
Pillow
uvicorn
pydantic